Impr: Move LerpingColor out of `lerp` package
